Refactoring of page to Pages
